+++ libs/angular-components/community/components/form/select/select.component.ts
@@ -9,7 +9,7 @@ import {
   ViewEncapsulation,
   ChangeDetectionStrategy,
   contentChildren,
-  AfterContentInit,
+  AfterContentChecked,
   OnInit,
   computed,
 } from "@angular/core";
@@ -56,8 +56,7 @@ import { ClosingButtonComponent } from "../../buttons/closing-button/closing-but
   },
 })
 export class SelectComponent
-  implements ControlValueAccessor, OnInit, AfterContentInit
-{
+  implements ControlValueAccessor, OnInit, AfterContentChecked {
   /**
    * The placeholder text to display when no option is selected.
    * @default ""
@@ -114,7 +113,7 @@ export class SelectComponent
     if (this.disabled()) this.setDisabledState(this.disabled());
   }
 
-  ngAfterContentInit() {
+  ngAfterContentChecked() {
     this.setDropdownWidth();
   }