+++ libs/react-components/src/community/helpers/hooks/use-breakpoint.ts
@@ -1,16 +1,39 @@
-import debounce from 'lodash-es/debounce';
-import { useEffect, useState } from 'react';
+import { debounce } from 'lodash-es';
+import { useLayoutEffect, useState } from 'react';
 
 export type Breakpoint = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
 export const breakpoints: Breakpoint[] = ['xs', 'sm', 'md', 'lg', 'xl', 'xxl'];
 
-export const useBreakpoint = (): Breakpoint | null => {
-  const [breakpoint, setBreakpoint] = useState(getBreakpoint());
+/**
+ * @param defaultServerBreakpoint - Default breakpoint for SSR, the hook returns this breakpoint in the SSR.
+ * @returns User's current breakpoint
+ */
+export const useBreakpoint = (defaultServerBreakpoint: Breakpoint = 'xs'): Breakpoint => {
+  const [breakpoint, setBreakpoint] = useState<Breakpoint>(defaultServerBreakpoint);
 
-  useEffect(() => {
+  useLayoutEffect(() => {
+    const getBreakpoint = (): Breakpoint => {
+      if (window.matchMedia('(min-width: 1400px)').matches) {
+        return 'xxl';
+      } else if (window.matchMedia('(min-width: 1200px)').matches) {
+        return 'xl';
+      } else if (window.matchMedia('(min-width: 992px)').matches) {
+        return 'lg';
+      } else if (window.matchMedia('(min-width: 768px)').matches) {
+        return 'md';
+      } else if (window.matchMedia('(min-width: 576px)').matches) {
+        return 'sm';
+      } else {
+        return 'xs';
+      }
+    };
+
     const resizeEvent = debounce((): void => {
       setBreakpoint(getBreakpoint());
     }, 20);
+
+    // Set the initial breakpoint on the client
+    setBreakpoint(getBreakpoint());
 
     window.addEventListener('resize', resizeEvent);
     return () => {
@@ -18,28 +41,6 @@ export const useBreakpoint = (): Breakpoint | null => {
       window.removeEventListener('resize', resizeEvent);
     };
   }, []);
-
-  return breakpoint;
-};
-
-const getBreakpoint = (): Breakpoint | null => {
-  let breakpoint: Breakpoint | null = null;
-
-  if (typeof window !== 'undefined') {
-    if (window.matchMedia('(min-width: 1400px)').matches) {
-      breakpoint = 'xxl';
-    } else if (window.matchMedia('(min-width: 1200px)').matches) {
-      breakpoint = 'xl';
-    } else if (window.matchMedia('(min-width: 992px)').matches) {
-      breakpoint = 'lg';
-    } else if (window.matchMedia('(min-width: 768px)').matches) {
-      breakpoint = 'md';
-    } else if (window.matchMedia('(min-width: 576px)').matches) {
-      breakpoint = 'sm';
-    } else {
-      breakpoint = 'xs';
-    }
-  }
 
   return breakpoint;
 };